Feature/add seeder score subtest3456789

+++ migrations/20200715080639-create-score-subtest.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       sw: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(4, 1)
       },
       umur: {
         type: Sequelize.INTEGER
